+++ mycodo/inputs/bmp280.py
@@ -140,10 +140,13 @@ class BMP280Sensor(AbstractInput):
         temperature = convert_units(
             'temperature', 'celsius', self.convert_to_unit,
             self.read_temperature())
-        pressure = self.read_pressure()
-        alt = altitude(pressure)
+        pressure_pa = self.read_pressure()
+        pressure = convert_units(
+            'pressure', 'pascals', self.convert_to_unit,
+            pressure_pa)
         alt = convert_units(
-            'altitude', 'meters', self.convert_to_unit, alt)
+            'altitude', 'meters', self.convert_to_unit,
+            altitude(pressure_pa))
         return temperature, pressure, alt
 
     def read(self):