+++ mycodo/tests/software_tests/factories/__init__.py
@@ -1,5 +1,3 @@
 # coding=utf-8
-from __future__ import absolute_import
-
 from .user_factory import UserFactory
 from .role_factory import RoleFactory